% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pointer_struct.R
\name{add_pointer_struct}
\alias{add_pointer_struct}
\title{Adds Pointer Structure to R objects for use in Pomp}
\usage{
add_pointer_struct(char, length)
}
\arguments{
\item{char}{A character vector.}

\item{length}{The length of the returned object.}
}
\value{
A character vector of the concatenated character string with sequential numbering
}
\description{
Adds Pointer Structure to R objects for use in Pomp
}
\examples{

add_pointer_struct("S", 3)
}
