% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_model.R
\name{aggregate_model}
\alias{aggregate_model}
\title{A Function to Apply Multiple Aggregations to Model Output}
\usage{
aggregate_model(df, aggregate_to = NULL, compartments = NULL,
  strat = NULL, hold_out_var = NULL, new_var = "incidence",
  total_pop = TRUE)
}
\arguments{
\item{df}{A dataframe of Model Output.}

\item{aggregate_to}{A character vector or list specifying the aggregation operations to perfrom on the
model output. Operations are carried out in the order specified. Implemented options are; disease, demographic,
and incidence.}

\item{compartments}{A character vector or list specifying the unique compartments to aggregate. May either be
specified once for all aggregation functions or for each function seperately.}

\item{strat}{The number of stratified groups in the model.}

\item{hold_out_var}{A character vector or list specifying the unique compartments that will not be aggregated. May either be
specified once for all aggregation functions or for each function seperately. If compartments is set then this arguement does
not need to be used.}

\item{new_var}{A character vevtor specifying the new variable to add when aggregating incidence.}

\item{total_pop}{A logical vector (defaults to \code{TRUE}) indicating if the total population should be
calculated when summarsing the model demographics.}
}
\value{
An aggregated dataframe.
}
\description{
A Function to Apply Multiple Aggregations to Model Output
}
\examples{

}
\seealso{
aggregate_model aggregate_model_internal combine_to_age_model combine_strat_model_output summarise_var_by_strat
}
