% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_model_internal.R
\name{aggregate_model_internal}
\alias{aggregate_model_internal}
\title{An Internal Function to Aggregate Model Output Using other Idmodelr functions.}
\usage{
aggregate_model_internal(df, aggregate_to, compartments, strat, hold_out_var,
  new_var = new_var, total_pop = TRUE)
}
\arguments{
\item{df}{A dataframe of Model Output.}

\item{aggregate_to}{A character vector specifying the aggregation function to apply possible values are;
disease, demographic, or incidence.}

\item{compartments}{A character vector specifying the unique compartments to aggregate.}

\item{strat}{The number of stratified groups in the model.}

\item{hold_out_var}{A character  vector specifying the unique compartments not to aggregate.}

\item{new_var}{A character vevtor specifying the new variable to add when aggregating incidence.}

\item{total_pop}{A logical vector (defaults to \code{TRUE}) indicating if the total population should be
calculated when summarsing the model demographics.}
}
\value{
An aggregated dataframe.
}
\description{
An Internal Function to Aggregate Model Output Using other Idmodelr functions.
}
\examples{

}
\seealso{
aggregate_model aggregate_model_internal combine_to_age_model combine_strat_model_output summarise_var_by_strat
}
