% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_polymod.R
\name{balance_polymod}
\alias{balance_polymod}
\title{Function to Balance a POLYMOD Like Matrix Using Population Data}
\usage{
balance_polymod(polymod, population, scale = NULL)
}
\arguments{
\item{polymod}{A POLYMOD like mixing matrix as produced by
\code{\link[idmodelr]{clean_polymod}} as a dataframe.}

\item{population}{A matching dataframe of populations by age_group,
with two variables: age_group and population.}

\item{scale}{A number to scale contacts by, defaults to null. With POLYMOD
style mixing matrices this gives contacts per day.}
}
\value{
A balanced POLYMOD style mixing matrix
}
\description{
Function to Balance a POLYMOD Like Matrix Using Population Data
}
\examples{

}
