% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_polymod.R
\name{clean_polymod}
\alias{clean_polymod}
\title{Clean a POLYMOD Like Mixing Matrix}
\usage{
clean_polymod(file_path)
}
\arguments{
\item{file_path}{A character string containing the path to the POLYMOD file}
}
\value{
A tidy data frame of POLYMOD like mixing data
}
\description{
Reads in and cleans a POLYMOD like mixing matrix saved as a .csv file.
POLYMOD mixing matrices are available from the orginal \href{https://doi.org/10.1371/journal.pmed.0050074}{paper}
}
\examples{

}
