% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_strat_model_output.R
\name{combine_strat_model_output}
\alias{combine_strat_model_output}
\title{Reduces the Dimensionality of a Stratifed Model}
\usage{
combine_strat_model_output(df, strat, compartments = c("S", "S_v", "E", "E_v",
  "I", "I_v", "N_SI", "N_SI_v", "N_EI", "N_EI_v"), hold_out_var = c("time",
  "traj"))
}
\arguments{
\item{df}{A data frame.}

\item{strat}{An integer specifying the number of stratifications to reduce.}

\item{compartments}{A character vector specifying the unique population compartments.}

\item{hold_out_var}{A character vector specifying the variables to keep unchanged.}
}
\value{
A dataframe of model output that has its dimensionality reduced
}
\description{
Reduces the Dimensionality of a Stratifed Model
}
\examples{

df <- data.frame(S1 = NA, S2 = NA, S3 = NA, time = NA)
combine_strat_model_output(df, 3, compartments = "S", hold_out_var = "time")

}
\seealso{
combine_to_age_model
}
