% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idmodelr.R
\docType{package}
\name{idmodelr-package}
\alias{idmodelr}
\alias{idmodelr-package}
\title{idmodelr: Utility Functions for Infectious Disease Modelling}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Contains utility functions for use in infectious disease modelling. Basic modelling tools are also supplied as starting blocks for more complex models.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.samabbott.co.uk/idmodelr}
  \item \url{https://github.com/seabbs/idmodelr}
  \item Report bugs at \url{https://github.com/seabbs/idmodelr/issues}
}

}
\author{
\strong{Maintainer}: Sam Abbott \email{contact@samabbott.co.uk} (0000-0001-8057-8037)

}
\keyword{internal}
