% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_plot.R
\name{interactive_plot}
\alias{interactive_plot}
\title{A Utility Function that Controls if a Plot is Returned/Printed and if it
is Interactive.}
\usage{
interactive_plot(plot, interactive = TRUE, print = TRUE, return = FALSE)
}
\arguments{
\item{plot}{A ggplot object}

\item{interactive}{A logicial indicating whether a interactive plot should be returned.
Defaults to \code{TRUE}}

\item{print}{A logical (defaults to \code{TRUE}) indicating if the plot should be printed.}

\item{return}{A logical (defaults to \code{FALSE}) indicating if the plot should be returned.}
}
\value{
A ggplot or ggplotly object.
}
\description{
A Utility Function that Controls if a Plot is Returned/Printed and if it
is Interactive.
}
\examples{


}
