% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_simulate.R
\name{model_simulate}
\alias{model_simulate}
\title{A Function to Simulate a Model from a Generic Simulation Function, with Pre and Post Processing}
\usage{
model_simulate(params, model, sim_fn, as_tibble = TRUE, ...)
}
\arguments{
\item{params}{A dataframe of parameters, with each parameter as a variable. Optionally a named vector can be used.}

\item{model}{A model compatible with your \code{sim_fn}. A \code{\link[pomp]{pomp}} model object is
recommended.}

\item{sim_fn}{A generic simulation function, with the first arguement as the model object,
a \code{params} arguement, and a \code{as.data.frame} arguement. Tested to work with \code{\link[pomp]{trajectory}}
  and \code{\link[pomp]{simulate}}.}

\item{as_tibble}{Logical (defaults to \code{TRUE}) indicating if the output
should be returned as a tibble, otherwise returned as the default \code{sim_fn} output.}

\item{...}{Additional arguments to pass to \code{sim_fn}}
}
\value{
Trajectories as a tibble, optionally returns the default \code{sim_fn} output.
}
\description{
A Function to Simulate a Model from a Generic Simulation Function, with Pre and Post Processing
}
\examples{

}
