% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_details.R
\docType{data}
\name{parameter_details}
\alias{parameter_details}
\title{Parameter Details}
\format{A data frame with 14 rows and 6 variables.
\describe{
\item{parameter}{Name of the parameter.}
\item{parameter_family}{Descriptive parameter family (i.e. transmission).}
\item{description}{What is this parameter.}
\item{type}{What type of parameter is this (i.e rate, proportion, probability etc.).}
\item{risk_stratified}{Is this parameter used in risk stratified models (yes/no).}
\item{non_exponential}{Is this parameter a non-exponential rate (yes/no).}
}}
\usage{
parameter_details
}
\description{
Details on the parameters used in models implemented in \code{idmodelr}.
}
\keyword{datasets}
