% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_comparision.R
\name{plot_model_comparision}
\alias{plot_model_comparision}
\title{Compare Deterministic Model Trajectories and Stochastic Mean Field Dynamics}
\usage{
plot_model_comparision(df1, df2, com_var, id_var = NULL, sum_fn = median,
  theme = theme_minimal, plot = TRUE, plot_titles = c("a.)", "b.)"),
  plots_as_list = FALSE, model_1 = NULL, model_2 = NULL)
}
\arguments{
\item{df1}{A dataframe containing the vector to be compared
for a model}

\item{df2}{A dataframe containing the vector to be compared for a second model}

\item{com_var}{The vector to be compared; use unquoted name (NSE).}

\item{id_var}{A character string indicating the id variable to summarise over
if required.}

\item{sum_fn}{The summary function to use to summarise the stochastic simulation, defaults to the median.}

\item{theme}{Supply a ggplot theme function to customise appearance.}

\item{plot}{Logical, if TRUE returns a plot,
if FALSE returns the intermediate dataframe.}

\item{plot_titles}{A length 2 character vector containing the plot titles, defaults to
a.), and b.).}

\item{plots_as_list}{Logical, if TRUE plots are returned as a list object, if FALSE (the default) plots are
combined using grid.arrange.}

\item{model_1}{A character string of the name to give the first model.}

\item{model_2}{A character string of the name to give the second model.}
}
\value{
A comparision plot between two model vectors
}
\description{
Compare Deterministic Model Trajectories and Stochastic Mean Field Dynamics
}
\examples{
library(dplyr)
## Example for unique model output
plot_model_comparision(iris, iris, Petal.Length)

## Example for simulated model output
df <- bind_rows(iris \%>\% mutate(sim = 1, id = 1:length(sim)),
 iris \%>\% mutate(sim = 2, id = 1:length(sim)))

 plot_model_comparision(df, iris, com_var = Petal.Length, id_var = "id")
}
\seealso{
plot_model_comparision model_df_to_vector aggregate_model_internal
}
