% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_comparision_internal.R
\name{plot_model_comparision_internal}
\alias{plot_model_comparision_internal}
\title{Compare a single vector of output from two models}
\usage{
plot_model_comparision_internal(model1_vector, model2_vector,
  theme = theme_minimal, plot = TRUE, plot_titles = c("a.)", "b.)"),
  plots_as_list = FALSE, model_1 = NULL, model_2 = NULL)
}
\arguments{
\item{model1_vector}{A numeric vector of model output.}

\item{model2_vector}{A numeric vector of model output.}

\item{theme}{Supply a ggplot theme function to customise appearance.}

\item{plot}{Logical, if TRUE returns a plot,
if FALSE returns the intermediate dataframe.}

\item{plot_titles}{A length 2 character vector containing the plot titles, defaults to
a.), and b.).}

\item{plots_as_list}{Logical, if TRUE plots are returned as a list object, if FALSE (the default) plots are
combined using grid.arrange.}

\item{model_1}{A character string of the name to give the first model.}

\item{model_2}{A character string of the name to give the second model.}
}
\value{
A comparision plot between two model vectors
}
\description{
Compare a single vector of output from two models
}
\examples{

plot_model_comparision_internal(c(1,2,3), c(1,2.1,3.2))

}
