% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_data}
\alias{save_data}
\title{A Function to Save Results in Multiple Formats}
\usage{
save_data(df, name = NULL, path = NULL, format = "rds", message = NULL,
  verbose = TRUE)
}
\arguments{
\item{df}{The data to be saved in any R supported format.}

\item{name}{A character string containing the name to save the data under.}

\item{path}{A character string containing the file pathway to the folder into
which to save the data. If not specified data will be saved into the home directory.}

\item{format}{A character vector specifying the format/formats to save the data into, defaults to rds. Currently
csv is also supported.}

\item{message}{An optional character string specifying the message to print.}

\item{verbose}{A logical indicating whether progress messages shoud be printed.}
}
\value{
Returns nothing, saves data in the specified file formats
}
\description{
A Function to Save Results in Multiple Formats
}
\examples{

## save_data(cars, name = "cars")
}
