% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIR_demo_ode.R
\name{SEIR_demo_ode}
\alias{SEIR_demo_ode}
\title{A Simple Susceptible-Exposed-Infected-Recovered Infectious Disease Model  with Simple Demographics}
\usage{
SEIR_demo_ode(t, x, params)
}
\arguments{
\item{t}{The timestep overwhich to calculate derivatives}

\item{x}{A numeric vector of compartment populations.}

\item{params}{A named vector of paramter values.}
}
\value{
A vector of derivatives
}
\description{
A Simple Susceptible-Exposed-Infected-Recovered Infectious Disease Model  with Simple Demographics
}
\examples{
##Model Input
S_0 <- 989
E_0 <- 10
I_0 <- 1
R_0 <- 0
beta <- 3
gamma <-  1/2
tau <- 2
mu <- 1/81
dt <- 1

parameters <- c(beta = beta, gamma = gamma, tau = tau, mu = mu)
inits <- c(S = S_0, E = E_0, I = I_0, R_0 = R_0)

SEIR_ode(1, inits, parameters)
}
