% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHLIR_risk_group_ode.R
\name{SHLIR_risk_group_ode}
\alias{SHLIR_risk_group_ode}
\title{A Simple Susceptible-High-risk-latent-Low-risk-latent-Infected-Recovered Infectious Disease Model with demographics}
\usage{
SHLIR_risk_group_ode(t, x, params)
}
\arguments{
\item{t}{The timestep overwhich to calculate derivatives}

\item{x}{A numeric vector of compartment populations.}

\item{params}{A named vector of paramter values.}
}
\value{
A vector of derivatives
}
\description{
A more complex SHLIR model flow diagram, with risk groups, treatment, and reinfection for those who have recovered from active disease
}
\examples{

## initialise
inits <- c(
# General population
S = 800,
H = 0,
L = 0,
I = 0,
Tr = 0,
R = 0,
## High risk population
S_H = 199,
H_H = 0,
L_H = 0,
I_H = 1,
Tr_H = 0,
R_H = 0
)

parameters <- c(
beta = 3, # Rate of transmission
beta_H = 6, # High risk rate of transmission
gamma_H = 1/5, # Rate of progression to active symptoms from high risk latent
nu = 1/2, #Rate of progression from high to low risk latent
gamma_L = 1/100, # Rate of progression to active symptoms for low risk latent
epsilon = 1/3, # Rate of treatment
tau = 1/2, # Rate of recovery
mu = 1/81, # Rate of natural mortality
p = 0.2, # proportion of new births that are high risk
M = 0.2 # Between group mixing
)

SHLIR_risk_group_ode(1, inits, parameters)

}
