% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI.R
\name{SI_demographics_ode}
\alias{SI_demographics_ode}
\title{Susceptible-Infected Model with Simple Demographics}
\usage{
SI_demographics_ode(t, x, params)
}
\arguments{
\item{t}{The timestep over which to calculate derivatives}

\item{x}{A numeric vector of compartment populations.}

\item{params}{A named vector of parameter values.}
}
\value{
A vector of derivatives
}
\description{
Susceptible-Infected Model with Simple Demographics
}
\examples{
##Model Input
S_0 <- 999
I_0 <- 1
beta <- 3
mu <- 1/81

parameters <- c(beta = beta, mu = mu)
inits <- c(S = S_0, I = I_0)

SI_demographics_ode(1, inits, parameters)
}
