% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_model.R
\name{simulate_model}
\alias{simulate_model}
\title{A Function to Simulate a Model from a Generic Simulation Function, with Pre and Post Processing}
\usage{
simulate_model(model, sim_fn, params, as_tibble = TRUE, aggregate_to = NULL,
  compartments = NULL, strat = NULL, hold_out_var = NULL,
  new_var = "incidence", total_pop = TRUE, summary_var, ...)
}
\arguments{
\item{model}{A model compatible with your \code{sim_fn}. A \code{\link[pomp]{pomp}} model object is
recommended.}

\item{sim_fn}{A generic simulation function, with the first arguement as the model object,
a \code{params} arguement, and a \code{as.data.frame} arguement. Tested to work with \code{trajectory}
and \code{simulate} from the \code{\link[pomp]{pomp}} package.}

\item{params}{A dataframe of parameters, with each parameter as a variable. Optionally a named vector can be used.}

\item{as_tibble}{Logical (defaults to \code{TRUE}) indicating if the output
should be returned as a tibble, otherwise returned as the default \code{sim_fn} output.}

\item{aggregate_to}{A character vector or list specifying the aggregation operations to perform on the
model output. Operations are carried out in the order specified. Implemented options are; disease, demographic,
and incidence.}

\item{compartments}{A character vector or list specifying the unique compartments to aggregate. May either be
specified once for all aggregation functions or for each function seperately.}

\item{strat}{The number of stratified groups in the model.}

\item{hold_out_var}{A character vector or list specifying the unique compartments that will not be aggregated. May either be
specified once for all aggregation functions or for each function seperately. If compartments is set then this arguement does
not need to be used.}

\item{new_var}{A character vector specifying the new variable to add when aggregating incidence.}

\item{total_pop}{A logical (defaults to \code{TRUE}) indicating if the total population should be
calculated when summarsing the model demographics.}

\item{summary_var}{A logical (defaults to \code{FALSE}), specifiying whether to add an additional summary variable across
all stratified levels.}

\item{...}{Additional arguments to pass to \code{sim_fn}}
}
\value{
Trajectories as a tibble, optionally returns the default \code{sim_fn} output.
}
\description{
A Function to Simulate a Model from a Generic Simulation Function, with Pre and Post Processing
}
\examples{

}
\seealso{
aggregate_model
}
