% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR_ode.R
\name{SIR_ode}
\alias{SIR_ode}
\title{A Simple Susceptible-Infected-Recovered Infectious Disease Model}
\usage{
SIR_ode(t, x, params)
}
\arguments{
\item{t}{The timestep overwhich to calculate derivatives}

\item{x}{A numeric vector of compartment populations.}

\item{params}{A named vector of paramter values.}
}
\value{
A vector of derivatives
}
\description{
A Simple Susceptible-Infected-Recovered Infectious Disease Model
}
\examples{
##Model Input
S_0 <- 989
I_0 <- 1
R_0 <- 0
beta <- 3
tau <- 2
dt <- 1

parameters <- c(beta = beta, tau = tau)
inits <- c(S = S_0, I = I_0, R_0 = R_0)

SIR_ode(1, inits, parameters)
}
