% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_model.R
\name{summarise_model}
\alias{summarise_model}
\title{Summarise Model Trajectories}
\usage{
summarise_model(traj = NULL, state.names = NULL, data = NULL,
  time.column = "time", summary = TRUE, replicate.column = "replicate",
  non.extinct = NULL, init.date = NULL, verbose = FALSE)
}
\arguments{
\item{traj}{Data frame, generic model output although tailored
to work with [pomp](https://www.rdocumentation.org/packages/pomp/versions/1.4.1.1/topics/pomp) model objects.}

\item{state.names}{A character vector. Names of the state variables to plot.
If \code{NULL} (default) all state variables are plotted.}

\item{data}{A data frame. Observation times and observed data. The time column must match that used in the model object, whilt
the observation name should also match one of those found in the model object.}

\item{time.column}{A character vector. The column in the data that indicates time}

\item{summary}{Logical. If \code{TRUE}, the mean, median as well as the 50th and 95th percentile
of the trajectories are plotted (default). If \code{FALSE}, all individual trajectories are plotted.}

\item{replicate.column}{A character vector. The column in the data that indicates the replicate
(if muliple replicates are to be plotted, i.e. if \code{summary} is \code{FALSE}}

\item{non.extinct}{A character vector. Names of the infected states which must be non-zero so the epidemic is still ongoing.
When the names of these states are provided, the extinction probability is plotted by computing the proportion of faded-out
epidemics over time. An epidemic has faded-out when all the infected states (whose names are provided) are equal to 0.
This is only relevant for stochastic models.}

\item{init.date}{Character. Date of the first point of the time series (default to \code{NULL}). If provided, the x-axis will be
in calendar format. NB: currently only works if the unit of time is the day.}

\item{verbose}{Logical (defaults to \code{FALSE}), indicates if internal process messages should be printed}
}
\value{
A list of summarised dataframes for an inputed model trajectory. These are the model trajectories, probability
  of extinction, and model summary trajectories.
}
\description{
This function checks, cleans and summarises inputed model trajectories. This function is based on
  [plotTraj](https://github.com/sbfnk/fitR/blob/master/R/plot.r) from
the [fitR](https://github.com/sbfnk/fitR) package written by [Sebaustian Funk](http://sbfnk.github.io/).
}
\examples{

}
\seealso{
plot_model
}
