% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pointer_struct.R
\name{add_pointer_struct}
\alias{add_pointer_struct}
\title{Adds Pointer Structure to R objects}
\usage{
add_pointer_struct(char, length)
}
\arguments{
\item{char}{A character vector.}

\item{length}{The length of the returned object.}
}
\value{
A character vector of the concatenated character string with sequential numbering
}
\description{
Adds Pointer Structure to R objects
}
\examples{

## For a single variable
add_pointer_struct("S", 3)

## For multiple variables
add_pointer_struct(c("S", "I"), length = 3)
}
