% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRS.R
\name{SEIRS_ode}
\alias{SEIRS_ode}
\title{Susceptible-Exposed-Infected-Recovered-Susceptible Model}
\usage{
SEIRS_ode(t, x, params)
}
\arguments{
\item{t}{The timestep overwhich to calculate derivatives}

\item{x}{A numeric vector of compartment populations.}

\item{params}{A named vector of paramter values.}
}
\value{
A vector of derivatives
}
\description{
Susceptible-Exposed-Infected-Recovered-Susceptible Model
}
\examples{
##Model Input
S_0 <- 989
E_0 <- 10
I_0 <- 1
R_0 <- 0
beta <- 3
gamma <-  1/2
tau <- 2
chi <- 0.5
dt <- 1

parameters <- c(beta = beta, gamma = gamma, tau = tau)
inits <- c(S = S_0, E = E_0, I = I_0, R_0 = R_0)

SEIRS_ode(1, inits, parameters)
}
