% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SI.R
\name{SI_ode}
\alias{SI_ode}
\title{Susceptible-Infected Model}
\usage{
SI_ode(t, x, params)
}
\arguments{
\item{t}{The timestep overwhich to calculate derivatives}

\item{x}{A numeric vector of compartment populations.}

\item{params}{A named vector of paramter values.}
}
\value{
A vector of derivatives
}
\description{
Susceptible-Infected Model
}
\examples{
##Model Input
S_0 <- 999
I_0 <- 1
beta <- 3
parameters <- c(beta = beta)
inits <- c(S = S_0, I = I_0)

SI_ode(1, inits, parameters)
}
