% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_model.R
\name{summarise_model}
\alias{summarise_model}
\title{Summarise Model Trajectories}
\usage{
summarise_model(traj = NULL, state.names = NULL, data = NULL,
  time.column = "time", summary = TRUE, replicate.column = "replicate",
  non.extinct = NULL, init.date = NULL, aggregate_to = NULL,
  compartments = NULL, strat = NULL, hold_out_var = NULL,
  new_var = "incidence", total_pop = TRUE, verbose = FALSE)
}
\arguments{
\item{traj}{Data frame, generic model output although tailored
to work with \href{https://www.rdocumentation.org/packages/pomp/versions/1.4.1.1/topics/pomp}{pomp} model objects.}

\item{state.names}{A character vector. Names of the state variables to plot.
If \code{NULL} (default) all state variables are plotted.}

\item{data}{A data frame. Observation times and observed data. The time column must match that used in the model object, whilt
the observation name should also match one of those found in the model object.}

\item{time.column}{A character vector. The column in the data that indicates time}

\item{summary}{Logical. If \code{TRUE}, the mean, median as well as the 50th and 95th percentile
of the trajectories are plotted (default). If \code{FALSE}, all individual trajectories are plotted.}

\item{replicate.column}{A character vector. The column in the data that indicates the replicate
(if muliple replicates are to be plotted, i.e. if \code{summary} is \code{FALSE}}

\item{non.extinct}{A character vector. Names of the infected states which must be non-zero so the epidemic is still ongoing.
When the names of these states are provided, the extinction probability is plotted by computing the proportion of faded-out
epidemics over time. An epidemic has faded-out when all the infected states (whose names are provided) are equal to 0.
This is only relevant for stochastic models.}

\item{init.date}{Character. Date of the first point of the time series (default to \code{NULL}). If provided, the x-axis will be
in calendar format. NB: currently only works if the unit of time is the day.}

\item{aggregate_to}{A character vector or list specifying the aggregation operations to perfrom on the
model output. Operations are carried out in the order specified. Implemented options are; disease, demographic,
and incidence.}

\item{compartments}{A character vector or list specifying the unique compartments to aggregate. May either be
specified once for all aggregation functions or for each function seperately.}

\item{strat}{The number of stratified groups in the model.}

\item{hold_out_var}{A character vector or list specifying the unique compartments that will not be aggregated. May either be
specified once for all aggregation functions or for each function seperately. If compartments is set then this arguement does
not need to be used.}

\item{new_var}{A character vevtor specifying the new variable to add when aggregating incidence.}

\item{total_pop}{A logical vector (defaults to \code{TRUE}) indicating if the total population should be
calculated when summarsing the model demographics.}

\item{verbose}{Logical (defaults to \code{FALSE}), indicates if internal process messages should be printed}
}
\value{
A list of summarised/aggregated dataframes for an inputed model trajectory. These are the model trajectories, probability
of extinction, and model summary trajectories.
}
\description{
This function checks, cleans, aggregates and summarises inputed model trajectories. This function is based on
\href{https://github.com/sbfnk/fitR/blob/master/R/plot.r}{plotTraj} from
the \href{https://github.com/sbfnk/fitR}{fitR} package written by \href{http://sbfnk.github.io/}{Sebaustian Funk}.
}
\examples{

}
\seealso{
plot_model aggregate_model
}
